refactor: Repository 인터페이스를 Port로 일원화
